LayoutTests/imported/w3c/web-platform-tests/css/css-tables/tentative/caption-expected.txt

Caption sizing and positioning

There is a fundamental disagreement between Chrome Legacy and Firefox on caption size influence on table width. Chrome uses caption min width as lower limit of table's grid box, FF does not. Firefox also does not support multiple captions.
Caption width limits

Caption minmax and grid width caption.min is lower limit of grid min.
auto
Caption.min vs caption.css_width Caption.css_width is upper limit of caption.min Caption.css_width is lower limit of caption.min
auto
auto
Caption width:50% Percent width is resolved wrt to table width.
auto
auto
Caption and size of empty table Larger of table borders, and caption min size.
cc
Caption height:80% github spec issue. TablesNG will not resolve percentage heights per spec. FF agrees.
auto
Multiple captions Do we allow multiple captions? There are 4 in this test.
over1
over2
under1
under2
2 above me, 2 below me?
Caption margins Margins between captions do not collapse
20px margins
1 caption
20px margins
20px margins
2 captions
Caption margins auto auto margins center captions that are less wide than the table.
auto margins
1 caption
auto margins
1 caption
Vertical writing mode caption
vertical caption bigger than table
vertical!

PASS table 1
PASS table 2
PASS table 3
PASS table 4
PASS table 5
PASS table 6
FAIL table 7 assert_equals:
<table style="width:200px;height:100px">
  <caption style="height:80%" data-expected-height="30"><div style="height:30px"></div></caption>
  <tbody><tr><td data-expected-width="200">auto</td>
</tr></tbody></table>
height expected 30 but got 80
PASS table 8
PASS table 9
PASS table 10
PASS table 11
PASS table 12
FAIL table 13 assert_equals:
<table>
  <caption style="writing-mode:vertical-rl;height:min-content;font:10px Ahem;color:rgba(0,0,0,0.3)" data-expected-height="100"><div style="height:100px;">vertical</div> caption bigger than table</caption>
  <tbody data-offset-y="100">
    <tr><td>vertical!</td>
  </tr></tbody>
</table>
offsetTop expected 100 but got 40
